feat(cpu): creation kernels (zeros/ones/full/arange/linspace) cover every advertised dtype
Audit Phase 2 Tasks 2.1+2.2: pre-fix the creation kernels handled only
Float32/Float64/Int32/Int64 and threw for everything else. This commit
extends coverage across all paths:

src/ops/creation.cpp (primary CPU path, inline):
- ones(): add UInt16/UInt32/UInt64/Int16 (were silently absent)
- full(float/double): add UInt16/UInt32/UInt64/Int16 to both overloads
- arange(): add UInt16/UInt32/UInt64 (Int8/Int16/UInt8/F16/BF16 already present)
- linspace(): add Int8/Int16/Int32/Int64 signed-integer dtypes

include/tenzor/ops/creation.hpp:
- full(double, ...) overload gains default dtype=Float64, device=cpu()
  so callers passing a double literal no longer silently narrow to float,
  fixing precision loss for Float64 and Int64 callers

src/backends/cpu/kernels/creation.cpp (secondary dispatch path):
- zeros_kernel: replace four-way switch with single std::memset (all
  IEEE/integer/complex/FP8/quant zeros are all-bits-zero)
- ones_kernel: extend to Float16/BFloat16/Int8/Int16/UInt8/UInt16/UInt32/
  UInt64/Bool/Complex64/Complex128/FP8_E4M3/FP8_E5M2
- full_kernel: widen value param float→double; add all missing dtypes
  matching ones_kernel coverage plus Int16/UInt16/UInt32/UInt64
- arange_kernel: widen start/end/step float→double; add Float16/BFloat16/
  Int8/Int16/UInt8/UInt16/UInt32/UInt64/Bool/Complex64/Complex128
- linspace_kernel: widen float→double; add Int8/Int16/Int32/Int64

src/backends/cpu/cpu_kernel_registry.cpp:
- Update forward declarations and lambda wrappers for full/arange/linspace
  to pass double instead of float (attrs.get_float() already returns double)

tests/unit/test_creation_dtype_coverage.cpp (new, 52 tests):
- zeros: Float16/BFloat16/Int8/Int16/UInt8/UInt16/UInt32/UInt64/Bool/
  Complex64/Complex128
- ones: same dtype matrix
- full: Float64 precision round-trip, Int64 no-truncation, plus Float16/
  BFloat16/Int8/Int16/UInt8/UInt16/UInt32/UInt64/Bool/Complex64/Complex128
- arange: Float16/BFloat16/Int8/Int16/UInt8/UInt16/UInt32/UInt64/Int64
  precision, Complex64/Complex128
- linspace: Float16/BFloat16/Int8/Int16/Int32/Int64

All 52/52 new tests pass. Regression sweep over 1378 matching tests shows
zero new failures (existing CUDA/ROCm/OneAPI Arange and Vulkan Transformer
failures are pre-existing and confirmed unchanged by stash comparison).

+++ src/backends/cpu/cpu_kernel_registry.cpp
@@ -398,12 +398,12 @@ namespace cpu {
     // Creation
     auto zeros_kernel(const std::vector<int64_t>& shape, DType dtype, const Device& device) -> Tensor;
     auto ones_kernel(const std::vector<int64_t>& shape, DType dtype, const Device& device) -> Tensor;
-    auto full_kernel(const std::vector<int64_t>& shape, float value, DType dtype, const Device& device) -> Tensor;
+    auto full_kernel(const std::vector<int64_t>& shape, double value, DType dtype, const Device& device) -> Tensor;
     auto rand_kernel(const std::vector<int64_t>& shape, DType dtype, const Device& device) -> Tensor;
     auto randn_kernel(const std::vector<int64_t>& shape, DType dtype, const Device& device) -> Tensor;
     auto randint_kernel(int64_t low, int64_t high, const std::vector<int64_t>& shape, DType dtype, const Device& device) -> Tensor;
-    auto arange_kernel(float start, float end, float step, DType dtype, const Device& device) -> Tensor;
-    auto linspace_kernel(float start, float end, int64_t steps, DType dtype, const Device& device) -> Tensor;
+    auto arange_kernel(double start, double end, double step, DType dtype, const Device& device) -> Tensor;
+    auto linspace_kernel(double start, double end, int64_t steps, DType dtype, const Device& device) -> Tensor;
     auto eye_kernel(int64_t n, int64_t m, DType dtype, const Device& device) -> Tensor;
 
     // RNN - Cell operations
@@ -3377,7 +3377,7 @@ void register_cpu_kernels(BackendDispatchTable& table) {
 
     table.register_kernel(OpId::Full, []([[maybe_unused]] std::span<const Tensor> inputs, const OpAttributes& attrs) {
         auto shape = attrs.get_int_list(AttrKey::Shape);
-        float value = static_cast<float>(attrs.get_float(AttrKey::Value, 0.0));
+        double value = attrs.get_float(AttrKey::Value, 0.0);
         int dtype_int = static_cast<int>(attrs.get_int(AttrKey::Dtype, static_cast<int64_t>(DType::Float32)));
         DType dtype = static_cast<DType>(dtype_int);
         Device device = Device::cpu();
@@ -3411,9 +3411,9 @@ void register_cpu_kernels(BackendDispatchTable& table) {
     });
 
     table.register_kernel(OpId::Arange, []([[maybe_unused]] std::span<const Tensor> inputs, const OpAttributes& attrs) {
-        float start = static_cast<float>(attrs.get_float(AttrKey::Start, 0.0));
-        float end = static_cast<float>(attrs.get_float(AttrKey::End, 0.0));
-        float step = static_cast<float>(attrs.get_float(AttrKey::Step, 1.0));
+        double start = attrs.get_float(AttrKey::Start, 0.0);
+        double end = attrs.get_float(AttrKey::End, 0.0);
+        double step = attrs.get_float(AttrKey::Step, 1.0);
         int dtype_int = static_cast<int>(attrs.get_int(AttrKey::Dtype, static_cast<int64_t>(DType::Float32)));
         DType dtype = static_cast<DType>(dtype_int);
         Device device = Device::cpu();
@@ -3421,8 +3421,8 @@ void register_cpu_kernels(BackendDispatchTable& table) {
     });
 
     table.register_kernel(OpId::Linspace, []([[maybe_unused]] std::span<const Tensor> inputs, const OpAttributes& attrs) {
-        float start = static_cast<float>(attrs.get_float(AttrKey::Start, 0.0));
-        float end = static_cast<float>(attrs.get_float(AttrKey::End, 1.0));
+        double start = attrs.get_float(AttrKey::Start, 0.0);
+        double end = attrs.get_float(AttrKey::End, 1.0);
         int64_t steps = attrs.get_int(AttrKey::Steps, 100);
         int dtype_int = static_cast<int>(attrs.get_int(AttrKey::Dtype, static_cast<int64_t>(DType::Float32)));
         DType dtype = static_cast<DType>(dtype_int);
